requirements.txt
numpy==1.15.2
scikit-image==0.14.0
torch==1.7.1
torchvision==0.2.1
pillow==7.2.0
tqdm
